feat: implement responsive design  landing page
