+++ apps/web/messages/en.json
@@ -2,7 +2,7 @@
   "Landing": {
     "student": "Student",
     "developer": "Developer",
-    "enthusiastic": "Enthusiastic"
+    "enthusiastic": "Enthusiast"
   },
   "Footer": {
     "contacts": "Let's keep in touch!",
