+++ libs/frontend/ui/landing/src/lib/components/background/background.module.scss
@@ -97,3 +97,22 @@
     transform: rotate(360deg);
   }
 }
+
+@media (max-width: theme('screens.sm')) {
+  .background {
+    .specialBlob {
+      height: 100px;
+      margin: -40px;
+
+      svg {
+        path {
+          stroke-width: 4;
+        }
+
+        text {
+          font-size: 10px;
+        }
+      }
+    }
+  }
+}
