+++ libs/frontend/ui/shared/src/index.ts
@@ -1,2 +1,3 @@
 export { FooterComponent } from './lib/components/footer';
 export { NavbarComponent } from './lib/components/navbar';
+export { ScreenTransitionComponent } from './lib/components/transitions/screen';
