+++ libs/frontend/ui/shared/src/lib/components/navbar/navbar.module.scss
@@ -165,3 +165,29 @@
     }
   }
 }
+
+@media (max-width: theme('screens.lg')) {
+  .navbar {
+    .intl {
+      width: 25vw;
+    }
+  }
+}
+
+@media (max-width: theme('screens.md')) {
+  .navbar {
+    justify-content: space-between;
+
+    ul {
+      display: none;
+    }
+  }
+}
+
+@media (max-width: theme('screens.sm')) {
+  .navbar {
+    .intl {
+      width: 30vw;
+    }
+  }
+}
